App/Server/requirements.txt
Flask
Flask-Cors
google-cloud-vision